Add support for reading single Enigma files
