+++ src/main/java/net/fabricmc/mappingio/MappingWriter.java
@@ -22,6 +22,7 @@ import java.io.Writer;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+import net.fabricmc.mappingio.format.EnigmaDirWriter;
 import net.fabricmc.mappingio.format.EnigmaWriter;
 import net.fabricmc.mappingio.format.MappingFormat;
 import net.fabricmc.mappingio.format.Tiny1Writer;
@@ -33,7 +34,7 @@ public interface MappingWriter extends Closeable, MappingVisitor {
 			return create(Files.newBufferedWriter(file), format);
 		} else {
 			switch (format) {
-			case ENIGMA: return new EnigmaWriter(file, true);
+			case ENIGMA_DIR: return new EnigmaDirWriter(file, true);
 			default: throw new UnsupportedOperationException("format "+format+" is not implemented");
 			}
 		}
@@ -45,6 +46,7 @@ public interface MappingWriter extends Closeable, MappingVisitor {
 		switch (format) {
 		case TINY: return new Tiny1Writer(writer);
 		case TINY_2: return new Tiny2Writer(writer, false);
+		case ENIGMA: return new EnigmaWriter(writer);
 		default: throw new UnsupportedOperationException("format "+format+" is not implemented");
 		}
 	}
